Move code to __init__.py
